fix(ci): use prysm and fix proof window command (#147)
Fixes error in the kurtosis e2e tests: 

```
  Caused by: Unexpected error occurred starting service 'el-1-reth-lighthouse'
  Caused by: An error occurred waiting for all TCP and UDP ports to be open for service 'el-1-reth-lighthouse' with private IP '172.16.0.10'; this is usually due to a misconfiguration in the service itself, so here are the logs:
  == SERVICE 'el-1-reth-lighthouse' LOGS ===================================
  error: unexpected argument '--rpc.eth-proof-window 100' found
  
    tip: a similar argument exists: '--rpc.eth-proof-window'
```


Transitions to pectra instead of setting at genesis because currently
kurtosis does not appear to work when enabling electra at genesis. The
deterministic deployer has to be set at genesis, so it's added as a
predeploy. This also waits for block 32, then runs the odyssey e2e
tests, and uses a pectra-compatible lighthouse image.