docs: update link RIP-7212 in `README.md` (#152)
Hi! I updated the link for `RIP-7212` in README.md because the new link
provides more specific and up-to-date information about `RIP-7212`. This
change will help developers find the necessary details more quickly and
also allow them, if needed, to navigate to the forum or other resources
directly from this
[page](https://github.com/ethereum/RIPs/blob/master/RIPS/rip-7212.md).

![image](https://github.com/user-attachments/assets/b95be33b-159e-4b7b-bc6d-472e2288aefd)